file cleanup, unified site readme and gh readme
