+++ .github/workflows/ci.yaml
@@ -1,4 +1,4 @@
-name: ci 
+name: Website CI
 on:
   push:
     branches:
