+++ README.md
@@ -1,4 +1,4 @@
-# IIIT Resources
+# IIITD Resources
 
 Contributors:
 
@@ -6,59 +6,23 @@ Contributors:
   <img src="https://contrib.rocks/image?repo=aflah02/SemWiseResourcesIIIT" />
 </a>
 
-If you want to add resources but confused about how to start, please refer to this [guide](https://github.com/aflah02/SemWiseResourcesIIIT/blob/main/CONTRIBUTING.md)
+The repository contains resources for IIITD courses. You can access the resources using the [website](https://aflah02.github.io/SemWiseResourcesIIIT/) or the README.md file.
 
-Note: The process has recently changed (December 2023) and the changes now need to go to `resources.yaml` and not to `README.md` directly
-
-[Essential Starters Guide](https://www.youtube.com/watch?v=dQw4w9WgXcQ)
-
-## Jump to specific semester
-- [Sem 1](#iiit-sem-1-resources)
-<center>
-
-| [IP](#ip) | [DC](#dc) | [LA](#la) | [HCI](#hci) | [COM](#com) | 
-|----|----|----|----|----|
-</center>
-
-- [Sem 2](#iiit-sem-2-resources)
-<center>
-
-| [DSA](#dsa) | [P&S](#ps) | [BE](#be) | [CO](#co) | 
-|----|----|----|----|
-</center>
-
-- [Sem 3](#iiit-sem-3-resources)
-<center>
-
-| [AP](#ap) | [OS](#os) | [DM](#dm) | [DS](#ds) | [M3](#m3) | [S&S](#ss) | [KCES](#kces) | 
-|----|----|----|----|----|----|----|
-</center>
-
-- [Sem 4](#iiit-sem-4-resources)
-<center>
-
-| [DBMS](#dbms) | [ITS](#its) | [ADA](#ada) | [M4](#m4) | [PIS](#pis) | [TOC](#toc) | [SML](#sml) | [PB](#pb) | 
-| ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | 
-
-| [GT](#gt) | 
-|----|
-</center>
-
-- [Sem 5-8](#iiit-3rd-and-4th-year-resources)
-<center>
-
-| [NLP](#nlp) | [ML](#ml) | [CN](#cn) | [NSS-1](#nss-1) | [RL](#rl) | [CG](#cg) | [EVS](#evs) | [ATP](#atp) | 
-| ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | 
-
-| [GPU](#gpu) | [DSCD](#dscd) | [VPM](#vpm) | [DL](#dl) | [IR](#ir) | [DSc](#dsc) | [WN](#wn) | [QM](#qm) | 
-| ---- | ---- | ---- | ---- | ---- | ---- | ---- | ---- | 
-
-| [GMT](#gmt) | [iROB](#irob) | [FCS](#fcs) | 
-|----|----|----|
-</center>
+- If you want to add resources but are confused about how to start, please refer to this [guide](https://aflah02.github.io/SemWiseResourcesIIIT/CONTRIBUTING)
+- Course books for courses can be found [here](https://drive.google.com/drive/folders/1Xhwlwbhj1HP6R9BysSoXcqScWFsnIj7B)
 
 
-## IIIT Sem 1 Resources
+> Note: The process has recently changed (December 2023) and the changes now need to go to `resources.yaml` and not to `README.md` directly
+
+| Semester | Courses |
+| --- | --- |
+| [Semester 1](semester-1) | [IP](semester-1#ip) [DC](semester-1#dc) [LA](semester-1#la) [HCI](semester-1#hci) [COM](semester-1#com) |
+| [Semester 2](semester-2) | [DSA](semester-2#dsa) [P&S](semester-2#ps) [BE](semester-2#be) [CO](semester-2#co) |
+| [Semester 3](semester-3) | [AP](semester-3#ap) [OS](semester-3#os) [DM](semester-3#dm) [DS](semester-3#ds) [M3](semester-3#m3) [S&S](semester-3#ss) [KCES](semester-3#kces) |
+| [Semester 4](semester-4) | [DBMS](semester-4#dbms) [ITS](semester-4#its) [ADA](semester-4#ada) [M4](semester-4#m4) [PIS](semester-4#pis) [TOC](semester-4#toc) [SML](semester-4#sml) [PB](semester-4#pb) [GT](semester-4#gt) |
+| [Open Electives](open-electives) | [NLP](open-electives#nlp) [ML](open-electives#ml) [CN](open-electives#cn) [NSS-1](open-electives#nss-1) [RL](open-electives#rl) [CG](open-electives#cg) [EVS](open-electives#evs) [ATP](open-electives#atp) [GPU](open-electives#gpu) [DSCD](open-electives#dscd) <br><br>[VPM](open-electives#vpm) [DL](open-electives#dl) [IR](open-electives#ir) [DSc](open-electives#dsc) [WN](open-electives#wn) [QM](open-electives#qm) [GMT](open-electives#gmt) [iROB](open-electives#irob) [FCS](open-electives#fcs) |
+
+## Semester 1
 ### IP
 - [Telusko Python Playlist](https://www.youtube.com/playlist?list=PLsyeobzWxl7poL9JTVyndKe62ieoN-MZ3)
 - [Code with Harry Python Playlist Hindi](https://www.youtube.com/playlist?list=PLu0W_9lII9agICnT8t4iYVSZ3eykIAOME)
@@ -126,7 +90,7 @@ Note: The process has recently changed (December 2023) and the changes now need 
 - [Aayi?](https://www.youtube.com/watch?v=y6120QOlsfU)
 
 
-## IIIT Sem 2 Resources
+## Semester 2
 ### DSA
 - [Java Tutorial Programming with Mosh](https://www.youtube.com/watch?v=eIrMbAQSU34)
 - [DSA Abdul Bari Sir](https://www.youtube.com/playlist?list=PLAXnLdrLnQpRcveZTtD644gM9uzYqJCwr)
@@ -158,7 +122,7 @@ Note: The process has recently changed (December 2023) and the changes now need 
 - [Playlist, Book and Slides by Sarangi (Official Course Book)](https://www.cse.iitd.ac.in/~srsarangi/archbooksoft.html)
 
 
-## IIIT Sem 3 Resources
+## Semester 3
 ### AP
 - [CS0150 Brown University](https://cs.brown.edu/courses/cs015/)
 - [Java Code with Harry Hindi (For Some OOPs Concepts)](https://www.youtube.com/playlist?list=PLu0W_9lII9agS67Uits0UnJyrYiXhDS6q)
@@ -205,7 +169,7 @@ Note: The process has recently changed (December 2023) and the changes now need 
 - [Aayi?](https://www.youtube.com/watch?v=y6120QOlsfU)
 
 
-## IIIT Sem 4 Resources
+## Semester 4
 ### DBMS
 - [Programming with Mosh Tutorial MySQL](https://www.youtube.com/watch?v=7S_tz1z_5bA)
 - [Knowledge GATE DBMS Playlist](https://www.youtube.com/playlist?list=PLmXKhU9FNesR1rSES7oLdJaNFgmuj0SYV)
@@ -251,7 +215,7 @@ Note: The process has recently changed (December 2023) and the changes now need 
 - Note, the course is not on algorithms
 
 
-## IIIT 3rd and 4th Year Resources
+## Open Electives
 ### NLP
 - [NPTEL Course](https://www.youtube.com/playlist?list=PLJJzI13YAXCHxbVgiFaSI88hj-mRSoMtI)
 - [Smoothing](https://www.youtube.com/watch?v=GwP8gKa-ij8)
@@ -276,7 +240,7 @@ Note: The process has recently changed (December 2023) and the changes now need 
 - A very good explanation of [Kerberos](https://www.fortinet.com/resources/cyberglossary/kerberos-authentication)
 - For other cryptographic concepts [Computerphile](https://www.youtube.com/@Computerphile)
 - For SSL/TLS, this [Cloudflare](https://blog.cloudflare.com/keyless-ssl-the-nitty-gritty-technical-details/) article, and [this](https://www.thesslstore.com/blog/explaining-ssl-handshake/) article for TLS1.2 and TLS1.3
-- IPSec - https://docs.strongswan.org/docs/5.9/howtos/ipsecProtocol.html
+- For IPSec, this [strongswan](https://docs.strongswan.org/docs/5.9/howtos/ipsecProtocol.html) article
 - Buffer Overflow and Format String Vulnerability - [Team bi0s wiki](https://wiki.bi0s.in/pwning/stack-overflow/intro/) and [This binary exploitation notes](https://ir0nstone.gitbook.io/notes/types/stack/introduction)
 - Lectures the biggest resource, right after asking the prof for doubts
 
@@ -359,9 +323,9 @@ Note: The process has recently changed (December 2023) and the changes now need 
 - MATLAB Simulink resources
 
 ### FCS
-- https://www.picoctf.org/
-- https://cryptohack.org/
-- https://prof.ninja/
+- [PicoCTF](https://www.picoctf.org)
+- [CryptoHack](https://cryptohack.org)
+- [Prof Ninja](https://prof.ninja)
 - Ofcourse, none of the above is "required" for the course and your proficiency is inversely proportional to the grade you might obtain
