+++ scripts/init_readme.md
